+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Ansible module `minio_alias` for managing `mc` aliases (#1)
+
 ## [0.0.1] - 2023-11-15
 
 ### Added